replaced type system by value system

+++ src/main/java/io/proleap/cobol/asg/metamodel/valuestmt/IntegerLiteralValueStmt.java
@@ -8,6 +8,8 @@
 
 package io.proleap.cobol.asg.metamodel.valuestmt;
 
+import java.math.BigDecimal;
+
 import io.proleap.cobol.asg.metamodel.IntegerLiteral;
 
 public interface IntegerLiteralValueStmt extends ValueStmt {
@@ -15,7 +17,7 @@ public interface IntegerLiteralValueStmt extends ValueStmt {
 	IntegerLiteral getLiteral();
 
 	@Override
-	Long getValue();
+	BigDecimal getValue();
 
 	void setIntegerLiteral(IntegerLiteral integerLiteral);
 }
